backend/requirements.txt
fastapi
pillow
qdrant-client
transformers
torch
